shl-assessment-rag/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
sentence-transformers
faiss-cpu
numpy
scikit-learn
